// src/utils/enums/typography-enums.ts
export enum ITypo {
	DISPLAY_LARGE = "display-large",
	DISPLAY_SMALL = "display-small",

	H1_TITLE = "H1-title",
	H2_TITLE = "H2-title",
	H3_TITLE = "H3-title",
	H4_TITLE = "H4-title",
	H5_TITLE = "H5-title",
	H6_TITLE = "H6-title",

	TEXTE_LG_SEMIBOLD = "text-lg-semibold",
	TEXTE_LG_MEDIUM = "text-lg-medium",
	TEXTE_LG_REGULAR = "text-lg-regular",

	TEXTE_MD_SEMIBOLD = "text-md-semibold",
	TEXTE_MD_MEDIUM = "text-md-medium",
	TEXTE_MD_REGULAR = "text-md-regular",

	TEXTE_SM_SEMIBOLD = "text-sm-semibold",
	TEXTE_SM_MEDIUM = "text-sm-medium",
	TEXTE_SM_REGULAR = "text-sm-regular",

	TEXTE_XS_SEMIBOLD = "text-xs-semibold",
	TEXTE_XS_MEDIUM = "text-xs-medium",
	TEXTE_XS_REGULAR = "text-xs-regular",
	

	OVERLINE_MEDIUM = "overline-sm",
	OVERLINE_SMALL = "overline-xs",
}

export enum ITypoColor {
	WHITE = "var(--color-generic-white)",
	BLACK = "var(--color-generic-black)",
	PRIMARY = "var(--text-primary)",
	SECONDARY = "var(--text-secondary)",
	INVERTED = "var(--text-inverted)",
}